change ptrdiff_t  alias for linux compatibility

+++ containers/list.hpp
@@ -6,7 +6,7 @@
 /*   By: llefranc <llefranc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/20 17:00:30 by llefranc          #+#    #+#             */
-/*   Updated: 2021/02/03 10:13:49 by llefranc         ###   ########.fr       */
+/*   Updated: 2021/02/03 10:52:51 by llefranc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ namespace ft
 
 			typedef	T											value_type;
 			typedef	Alloc										allocator_type;
-			typedef	ptrdiff_t									difference_type;
+			typedef	long long int								difference_type;
 			typedef	size_t										size_type;
 
 			typedef	T&											reference;
